Moved HttpServletRequest wrapper to xsodata utils
